Add fetchAndCache to core/utils Fix a few nits
